requirements.txt
typer>=0.9.0
rich>=13.7.0
openai>=1.12.0
python-dotenv>=1.0.0
